fix rename file
